Fixed bugs chatid and added vizualization of measures

+++ src/main/java/lifestylecoach/client/models/User.java
@@ -2,11 +2,11 @@ package lifestylecoach.client.models;
 
 public class User {
 
-    public Integer uid;
+    public Long uid;
     public String name;
     public String surname;
 
-    public User(Integer uid, String name, String surname) {
+    public User(Long uid, String name, String surname) {
         this.uid = uid;
         this.name = name;
         this.surname = surname;
